requirements.txt
sanic
transformers
accelerate
rembg
aiohttp==3.8.1
asyncer==0.0.1
click==8.1.3
fastapi==0.80.0
filetype==1.1.0
gdown==4.5.1
imagehash==4.2.1
numpy==1.21.6
onnxruntime==1.12.1
opencv-python-headless==4.6.0.66
pillow==9.2.0
pymatting==1.1.8
python-multipart==0.0.5
scikit-image==0.19.3
scipy==1.7.3
tqdm==4.64.0
uvicorn==0.18.3
watchdog==2.1.9
